Pieced together all the logic:
a. A Server that accepts client connections.
b. A Client connection class that receives messages and fires an async event so the Aggregator can process the number and update the stats.
c. Aggregator process the input / message and fires a write to file event if need be.
d. The file writer writes the number to /tmp/numbers.log file.
e. Various tests to verify the functionality.
f. Also included performance test class.